fix: address review points — correct file path, add GetCustomFieldsTest, use factories
- Move test file from esp/program/controllers/ to
  esp/program/modules/tests/ per proposal spec (refs #3780, #3773)
- Add GetCustomFieldsTest using CacheFlushTestCase as committed in
  proposal: tests empty tag, nonexistent form class, empty JSON list
- Refactor _make_saved_class() to use make_class() factory instead of
  ClassSubject.objects.create() directly
- Clarify is-not regression guard docstring: fix is already in place,
  test documents correct behaviour and prevents future regression
- Remove AttachAndAssociateTest (out of proposal scope for this PR)
